refactor: projection 패키지 변경
