+++ src/main/java/be/dash/dashserver/database/core/lesson/LessonJpaEntityRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import be.dash.dashserver.core.domain.common.Genre;
-import be.dash.dashserver.core.domain.teacher.projection.TeacherLessonCount;
+import be.dash.dashserver.database.core.teacher.projection.TeacherLessonCount;
 
 public interface LessonJpaEntityRepository extends JpaRepository<LessonJpaEntity, Long>, JpaSpecificationExecutor<LessonJpaEntity> {
 
-    @Query("select new be.dash.dashserver.core.domain.teacher.projection.TeacherLessonCount(t.id, t.member.nickname, count(l)) " +
+    @Query("select new be.dash.dashserver.database.core.teacher.projection.TeacherLessonCount(t.id, t.member.nickname, count(l)) " +
             "from LessonJpaEntity l " +
             "join l.teacher t " +
             "group by t.member.nickname " +
