+++ src/main/java/be/dash/dashserver/database/core/teacher/projection/TeacherLessonCount.java
@@ -1,4 +1,4 @@
-package be.dash.dashserver.core.domain.teacher.projection;
+package be.dash.dashserver.database.core.teacher.projection;
 
 public record TeacherLessonCount(Long teacherId, String nickname, Long lessonCount) {
 }
